[connectors] Rework loader CVE + IPInfo
