+++ manifest.json
@@ -28,8 +28,10 @@
       "subscription_link": "https://nvd.nist.gov/contact",
       "source_code": "https://github.com/OpenCTI-Platform/connectors/tree/master/external-import/cve",
       "type": "object",
-      "additionalProperties": true,
       "default": {
+        "CONNECTOR_ID": "cve--e0c380ad-6665-4f2e-8558-5d2610e5abcf",
+        "CONNECTOR_NAME": "Common Vulnerabilities and Exposures",
+        "CONNECTOR_SCOPE": "cve",
         "CONNECTOR_LOG_LEVEL": "error",
         "CONNECTOR_DURATION_PERIOD": "PT24H",
         "CVE_BASE_URL": "https://services.nvd.nist.gov/rest/json/cves",
@@ -43,6 +45,18 @@
         "CVE_API_KEY"
       ],
       "properties": {
+        "CONNECTOR_ID": {
+          "type": "string",
+          "description": "A unique UUIDv4 identifier for this connector instance."
+        },
+        "CONNECTOR_NAME": {
+          "type": "string",
+          "description": "Name of the connector."
+        },
+        "CONNECTOR_SCOPE": {
+          "type": "string",
+          "description": "The scope or type of data the connector is importing, either a MIME type or Stix Object (for information only)."
+        },
         "CONNECTOR_LOG_LEVEL": {
           "description": "Determines the verbosity of the logs.",
           "enum": [
@@ -56,18 +70,6 @@
         "CONNECTOR_DURATION_PERIOD": {
           "description": "Duration between two scheduled runs of the connector (ISO 8601 format).",
           "format": "duration",
-          "type": "string"
-        },
-        "CONNECTOR_ID": {
-          "description": "A unique UUIDv4 identifier for this connector instance.",
-          "type": "string"
-        },
-        "CONNECTOR_NAME": {
-          "description": "Name of the connector.",
-          "type": "string"
-        },
-        "CONNECTOR_SCOPE": {
-          "description": "The scope or type of data the connector is importing, either a MIME type or Stix Object (for information only).",
           "type": "string"
         },
         "CONNECTOR_QUEUE_THRESHOLD": {
@@ -328,9 +330,9 @@
       "subscription_link": "https://ipinfo.io/support",
       "source_code": "https://github.com/OpenCTI-Platform/connectors/tree/master/internal-enrichment/ipinfo",
       "type": "object",
-      "additionalProperties": true,
       "default": {
-        "CONNECTOR_NAME": "IPInfo",
+        "CONNECTOR_ID": "ipinfo--43cf861c-72a7-4e45-864a-b19e32e6a8bc",
+        "CONNECTOR_NAME": "IP Info",
         "CONNECTOR_SCOPE": "IPv4-Addr,IPv6-Addr",
         "CONNECTOR_LOG_LEVEL": "error",
         "CONNECTOR_AUTO": true,
@@ -338,35 +340,34 @@
         "IPINFO_USE_ASN_NAME": true
       },
       "required": [
-        "CONNECTOR_ID",
         "IPINFO_TOKEN"
       ],
       "properties": {
+        "CONNECTOR_ID": {
+          "type": "string",
+          "description": "A unique UUIDv4 identifier for this connector instance."
+        },
         "CONNECTOR_NAME": {
-          "description": "Name of the connector.",
-          "type": "string"
+          "type": "string",
+          "description": "Name of the connector."
         },
         "CONNECTOR_SCOPE": {
-          "description": "The scope or type of data the connector is importing, either a MIME type or Stix Object (for information only).",
-          "type": "string"
+          "type": "string",
+          "description": "The scope or type of data the connector is importing, either a MIME type or Stix Object (for information only)."
         },
         "CONNECTOR_LOG_LEVEL": {
           "description": "Determines the verbosity of the logs.",
-          "type": "string",
           "enum": [
             "debug",
             "info",
             "warn",
             "error"
-          ]
+          ],
+          "type": "string"
         },
         "CONNECTOR_AUTO": {
           "description": "Enables or disables automatic enrichment of observables for OpenCTI.",
           "type": "boolean"
-        },
-        "CONNECTOR_ID": {
-          "type": "string",
-          "description": "A unique UUIDv4 identifier for this connector instance."
         },
         "CONNECTOR_LISTEN_PROTOCOL": {
           "description": "Protocol used for listening.",
@@ -398,14 +399,14 @@
         },
         "IPINFO_MAX_TLP": {
           "description": "Traffic Light Protocol (TLP) level to apply on objects imported into OpenCTI.",
-          "type": "string",
           "enum": [
             "clear",
             "green",
             "amber",
             "amber+strict",
             "red"
-          ]
+          ],
+          "type": "string"
         },
         "IPINFO_USE_ASN_NAME": {
           "description": "If enabled, uses the ASN name instead of the ASN number in enrichment results.",
